requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.12.1
packaging==16.8
protobuf==3.2.0
pyparsing==2.2.0
six==1.10.0
tensorflow-gpu==1.0.1
Werkzeug==0.12.1
